backend/requirements.txt
Django==4.2
djoser==2.3.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
requests==2.32.4
django-cors-headers==4.7.0
gunicorn==23.0.0
